Revamp
restyle the code

add shell compile

update the copyright comment

add const attribute to FN_ functions; uninline _addbits

extra check ctx->len in sha256_hash()

update Makefile

update self test

add assumptions for cbmc

get rid of MINIMIZE_STACK_IMPACT, move W to a context

replace legacy with Mark 2 as the only variant

update the README file

update copyright
