+++ LICENSE
@@ -1,6 +1,8 @@
 MIT License
 
-Copyright (c) 2021 Ilya O. Levin
+Copyright (c) 2010,2014 Literatecode, http://www.literatecode.com
+Copyright (c) 2022 Ilia Levin (ilia@levin.sg)
+
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
